requirements-dev.txt
# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.18.1

# Code quality
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1
django-stubs==4.2.7

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0